requirements_clip_server.txt
fastapi==0.101.1
pydantic
uvicorn
pymongo==4.5.0
python-dotenv
Pillow==9.0.0
minio
Requests==2.30.0
passlib
python-jose
python-multipart
transformers==4.29.2
safetensors==0.3.1
openai-clip
Pillow==9.0.0
pytorch_lightning
torchvision==0.15.2
torch==2.0.1
torchinfo==1.8.0
opencv-python==4.8.0.74
blendmodes
scikit-image
lark
diffusers
termcolor
msgpack